Final code review and cleaning
