+++ Navigation/AppState/Constants.swift
@@ -15,10 +15,6 @@ struct Constants {
 }
 
 public extension UIView {
-
-    func toAutoLayout() {
-        translatesAutoresizingMaskIntoConstraints = false
-    }
     
     func addSubviews(_ subviews: UIView...) {
           subviews.forEach { addSubview($0) }
