+++ Navigation/AppState/PhotosArray.swift
